# server/openapi.yaml
openapi: 3.0.0
info:
  title: API Roadtrips
  version: 1.0.0
  description: Documentation de l'API pour la gestion des roadtrips
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Roadtrip:
      type: object
      properties:
        id:
          type: string
          example: '12345'
        name:
          type: string
          example: Mon Roadtrip
    RoadtripResponse:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        name:
          type: string
        startLocation:
          type: string
        startDateTime:
          type: string
          format: date-time
        endLocation:
          type: string
        endDateTime:
          type: string
          format: date-time
        currency:
          type: string
        notes:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/File'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/File'
        thumbnail:
          $ref: '#/components/schemas/File'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        __v:
          type: integer
    Step:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        roadtripId:
          type: string
        type:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        arrivalDateTime:
          type: string
          format: date-time
        departureDateTime:
          type: string
          format: date-time
        travelTime:
          type: number
        isArrivalTimeConsistent:
          type: boolean
        notes:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/File'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/File'
        thumbnail:
          $ref: '#/components/schemas/File'
        accommodations:
          type: array
          items:
            $ref: '#/components/schemas/Accommodation'
        activities:
          type: array
          items:
            $ref: '#/components/schemas/Activity'
        __v:
          type: integer
    Accommodation:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        stepId:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        website:
          type: string
        phone:
          type: string
        email:
          type: string
        reservationNumber:
          type: string
        confirmationDateTime:
          type: string
          format: date-time
        arrivalDateTime:
          type: string
          format: date-time
        departureDateTime:
          type: string
          format: date-time
        nights:
          type: integer
        price:
          type: number
        notes:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/File'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/File'
        thumbnail:
          $ref: '#/components/schemas/File'
        __v:
          type: integer
    Activity:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        stepId:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        website:
          type: string
        phone:
          type: string
        email:
          type: string
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        duration:
          type: number
        typeDuration:
          type: string
        reservationNumber:
          type: string
        price:
          type: number
        notes:
          type: string
        photos:
          type: array
          items:
            $ref: '#/components/schemas/File'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/File'
        thumbnail:
          $ref: '#/components/schemas/File'
        __v:
          type: integer
    File:
      type: object
      properties:
        _id:
          type: string
        url:
          type: string
        type:
          type: string
        fileId:
          type: string
        createdAt:
          type: string
          format: date-time
        __v:
          type: integer
paths:
  /:
    post:
      summary: Créer un nouveau roadtrip
      tags:
        - Roadtrips
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                startLocation:
                  type: string
                startDateTime:
                  type: string
                  format: date-time
                endLocation:
                  type: string
                endDateTime:
                  type: string
                  format: date-time
                currency:
                  type: string
                notes:
                  type: string
                thumbnail:
                  type: string
                  format: binary
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
                documents:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Roadtrip créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoadtripResponse'
        '400':
          description: Requête invalide
        '401':
          description: Non autorisé
        '500':
          description: Erreur interne du serveur
  /{idRoadtrip}/steps:
    post:
      summary: Créer une nouvelle étape pour un roadtrip
      tags:
        - Steps
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idRoadtrip
          required: true
          schema:
            type: string
          description: ID du roadtrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                location:
                  type: string
      responses:
        '201':
          description: Étape créée avec succès
        '400':
          description: Requête invalide
        '401':
          description: Non autorisé
        '500':
          description: Erreur interne du serveur
  /{idRoadtrip}/steps/{idStep}/accommodations:
    post:
      summary: Créer un nouvel hébergement pour une étape de roadtrip
      tags:
        - Accommodations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idRoadtrip
          required: true
          schema:
            type: string
          description: ID du roadtrip
        - in: path
          name: idStep
          required: true
          schema:
            type: string
          description: ID de l'étape
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  format: binary
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
                documents:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Hébergement créé avec succès
        '400':
          description: Requête invalide
        '401':
          description: Non autorisé
        '500':
          description: Erreur interne du serveur
  /{idRoadtrip}/steps/{idStep}/activities:
    post:
      summary: Créer une nouvelle activité pour une étape de roadtrip
      tags:
        - Activities
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: idRoadtrip
          required: true
          schema:
            type: string
          description: ID du roadtrip
        - in: path
          name: idStep
          required: true
          schema:
            type: string
          description: ID de l'étape
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  format: binary
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
                documents:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Activité créée avec succès
        '400':
          description: Requête invalide
        '401':
          description: Non autorisé
        '500':
          description: Erreur interne du serveur
tags: []